+++ apps/nextjs-website/src/lib/strapi/tutorial.ts
@@ -9,11 +9,12 @@ import { BlocksContentCodec } from './codecs/BlocksContentCodec';
 import { PaginationCodec } from './codecs/PaginationCodec';
 import { ProductCodec } from './codecs/ProductCodec';
 import { RelatedLinksCodec } from './codecs/RelatedLinksCodec';
+import { PartCodec } from './codecs/PartCodec';
 
 const BannerLinkCodec = t.strict({
   id: t.number,
   title: t.union([NullToUndefinedCodec, t.string]),
-  body: t.union([NullToUndefinedCodec, BlocksContentCodec]),
+  content: t.union([NullToUndefinedCodec, BlocksContentCodec]),
 });
 
 export const TutorialCodec = t.strict({
@@ -21,7 +22,7 @@ export const TutorialCodec = t.strict({
   attributes: t.strict({
     title: t.string,
     slug: t.string,
-    content: BlocksContentCodec,
+    parts: t.array(PartCodec),
     createdAt: DateFromISOString,
     updatedAt: DateFromISOString,
     publishedAt: t.union([NullToUndefinedCodec, tt.DateFromISOString]),
@@ -49,8 +50,11 @@ const makeStrapiTutorialsPopulate = () =>
       image: {
         populate: ['image'],
       },
+      parts: '*',
       product: { populate: 'logo' },
-      bannerLinks: '*',
+      bannerLinks: {
+        populate: ['icon'],
+      },
     },
   });