+++ apps/storybook-app/stories/molecules/Chat.stories.tsx
@@ -1,6 +1,7 @@
 import { Decorator, Meta, StoryObj } from '@storybook/react';
 import Chat from '../../../nextjs-website/src/components/molecules/Chat/Chat';
 import React from 'react';
+import { nextIntlContextDecorator } from '../next-intl-context.helper';
 
 const meta: Meta<typeof Chat> = {
   title: 'Molecules/Chat',
@@ -8,9 +9,7 @@ const meta: Meta<typeof Chat> = {
 };
 
 const decorator: Decorator = (story) => (
-  <div style={{ backgroundColor: 'black', padding: '2rem' }}>
-    {story()}
-  </div>
+  <div style={{ backgroundColor: 'black', padding: '2rem' }}>{story()}</div>
 );
 
 export default meta;
@@ -19,24 +18,26 @@ export const Showcase: StoryObj<typeof Chat> = {
   args: {
     chatMessages: [
       {
-        message: "Ciao, sono [Nome Chatbot], il chatbot di DevPortal.\nPrima di incominciare ti invito a leggere l'Informativa sulla Privacy.\nScrivi Accetto per proseguire.",
-        timestamp: '11:21'
+        message:
+          "Ciao, sono [Nome Chatbot], il chatbot di DevPortal.\nPrima di incominciare ti invito a leggere l'Informativa sulla Privacy.\nScrivi Accetto per proseguire.",
+        timestamp: '11:21',
       },
       {
-        message: 'Accetto', 
-        sender: 'Mario Rossi', 
-        timestamp: '11:22'
-      },
-      {
-        message: "Ottimo! Adesso possiamo incominciare. Come posso aiutarti?",
-        timestamp: '11:22'
-      },
-      {
-        message: "E' possibile pagare una posizione debitoria una volta scaduta?",
+        message: 'Accetto',
         sender: 'Mario Rossi',
-        timestamp: '11:22'
-      }
-    ]
+        timestamp: '11:22',
+      },
+      {
+        message: 'Ottimo! Adesso possiamo incominciare. Come posso aiutarti?',
+        timestamp: '11:22',
+      },
+      {
+        message:
+          "E' possibile pagare una posizione debitoria una volta scaduta?",
+        sender: 'Mario Rossi',
+        timestamp: '11:22',
+      },
+    ],
   },
-  decorators: [decorator],
+  decorators: [decorator, nextIntlContextDecorator],
 };